pyproject.toml
[project]
name = "repl-nix-yadiaaihub"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "openai>=1.57.2",
    "sqlalchemy>=2.0.36",
    "psycopg2-binary",
    "email-validator",
    "flask-session>=0.8.0",
    "firebase-admin>=6.6.0",
    "python-slugify>=8.0.4",
]
